requirements.txt
pyFirmata>=2-2.4.2